Code/requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
backports.weakref==1.0rc1
bleach==1.5.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
dnspython==1.16.0
eventlet==0.28.0
Flask==1.1.2
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
isort==5.5.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Keras==1.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==2.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
mccabe==0.6.1
monotonic==1.5
multidict==4.7.6
numpy==1.18.5
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opt-einsum==3.3.0
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.2.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
Theano==1.0.5
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
zipp==3.2.0
